refactor: move sampling to a different package; reorganize sampling tests
